+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ CHANGELOG
 1.3.dev1
 ========
 
+* bug-fix: Remove __all__ and add noqa in __init__
 * bug-fix: Estimators: Change max_iterations hyperparameter key for KMeans
 
 1.3.0