+++ CHANGELOG.rst
@@ -5,8 +5,11 @@ CHANGELOG
 1.3.dev1
 ========
 
+* bug-fix: Remove __all__ and add noqa in __init__
 * bug-fix: Estimators: Change max_iterations hyperparameter key for KMeans
 * bug-fix: Unit Tests: Improve unit test runtime
+* bug-fix: Estimators: Remove unused argument job_details for ``EstimatorBase.attach()``
+
 
 1.3.0
 =====